Update patch set 2
Patch Set 2: Code-Review+1

tvdpi: Resources for screens somewhere between mdpi and hdpi; approximately 213dpi. This is not considered a "primary" density group. It is mostly intended for televisions and most apps shouldn't need it—providing mdpi and hdpi resources is sufficient for most apps and the system will scale them as appropriate.

Patch-set: 2
Label: Code-Review=+1